backend/app/serializers.py
from rest_framework import serializers

from .models import MediaModel


class MediaSerializer(serializers.ModelSerializer):
    image = serializers.ImageField(use_url=True)

    class Meta:
        model = MediaModel
        fields = ["title", "image", "thumbnail"]

    def get_image(self, obj):
        request = self.context.get("request")
        if obj.image and request:
            return request.build_absolute_uri(obj.image.url)
        return None
